% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameSpace.R
\name{nameSpace_addBatch}
\alias{nameSpace_addBatch}
\title{mediawiki域名空间批量增加}
\usage{
nameSpace_addBatch(data, file = "./res.txt")
}
\arguments{
\item{data}{配置文件，文件列号必须为NS_Name,NS_Title,NS_Id}

\item{file}{结果存放名称}
}
\value{
无返回值
}
\description{
mediawiki域名空间批量增加
}
\examples{
nameSpace_addBatch();
}
