% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameSpace.R
\name{nameSpace_dataTemplate}
\alias{nameSpace_dataTemplate}
\title{生成nameSpace批量新增的配置文件}
\usage{
nameSpace_dataTemplate()
}
\value{
返回df的数据集
}
\description{
生成nameSpace批量新增的配置文件
}
\examples{
nameSpace_dataTemplate();
}
