% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameSpace.R
\name{nameSpace_demoBatch}
\alias{nameSpace_demoBatch}
\title{命名空间演示数据处理}
\usage{
nameSpace_demoBatch(file = "./res.txt")
}
\arguments{
\item{file}{结算处理文件名}
}
\value{
返回值
}
\description{
命名空间演示数据处理
}
\examples{
nameSpace_demoBatch("./res.txt");
}
