% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nameSpace.R
\name{ns_cat}
\alias{ns_cat}
\title{优化cat,用于输入到文档}
\usage{
ns_cat(x, file)
}
\arguments{
\item{x}{R对象}

\item{file}{目前文件}
}
\value{
直接输出到文件，没有返回值
}
\description{
优化cat,用于输入到文档
}
\examples{
ns_cat('hello world',"./helloworld.txt");
}
