% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmTable.R
\name{write.wikiTable}
\alias{write.wikiTable}
\title{将数据框转化为wikiTable可以识别的格式}
\usage{
write.wikiTable(df = iris, sortable = TRUE, file = "./wikiTable.txt")
}
\arguments{
\item{df}{数据框}

\item{sortable}{是否排序}

\item{file}{输出文件目录}
}
\value{
返回值
}
\description{
将数据框转化为wikiTable可以识别的格式
}
\examples{
write.wikiTable();
}
